requirements.txt
singledispatch==3.4.0.3
pycryptodome==3.8.2
uwsgi == 2.0.18
Django == 2.1.3
django-cors-headers == 2.4.0
djangorestframework == 3.8.2
HttpRunner == 1.5.15
mysqlclient == 1.3.13
beautifulsoup4 == 4.6.3
tornado == 5.1.1
django-celery == 3.2.2
rsa==4.0
PyMySQL==0.9.3
